UPLOAD_TO_PYTHONANYWHERE/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
lxml==4.9.3